+++ packages/http-client/src/lib/clients/alerts/AlertsClient.types.ts
@@ -1,3 +1,4 @@
+import { AxiosInstance } from 'axios'
 import { Parser } from '../../core/http-service/HttpService.types'
 import { Newable } from '../../util.types'
 
@@ -38,28 +39,29 @@ export type listObjectResponse = {
 }
 
 export type DetectionItem = {
-  aid: string;
-  oid: string;
-  tid: string;
-  mjd: number;
-  candid: string;
-  fid: number;
-  isdiffpos: number;
-  mag: number;
-  e_mag: number;
-  ra: number;
-  dec: number;
-  rb: number;
-  rbversion: string;
-  has_stamp: boolean;
-  corrected: boolean;
-  step_id_corr: string;
-  parent_candid: string;
+  aid: string
+  oid: string
+  tid: string
+  mjd: number
+  candid: string
+  fid: number
+  isdiffpos: number
+  mag: number
+  e_mag: number
+  ra: number
+  dec: number
+  rb: number
+  rbversion: string
+  has_stamp: boolean
+  corrected: boolean
+  step_id_corr: string
+  parent_candid: string
 }
 
 export interface singleObjectResponse extends objectListItem {}
 
 export interface IAlertsClient {
+  initClient(axiosInstance?: AxiosInstance): void
   queryObjects<T>(
     objectFilters: ObjectFilters,
     parser?: Parser<listObjectResponse, T>,