+++ packages/http-client/src/lib/core/http-service/HttpService.ts
@@ -29,14 +29,13 @@ export class HttpService implements IHttpService {
    * @param axiosInstance - optional axios instance to use instead of creating a new one
    * @param accessToken - token for authentication with the API
    */
-  connect(baseUrl: string, axiosInstance?: AxiosInstance, accessToken = '') {
+  initClient(baseUrl: string, axiosInstance?: AxiosInstance, accessToken = '') {
     if (axiosInstance) {
       this.axiosInstance = axiosInstance
     } else {
       this.axiosInstance = axios.create({
         baseURL: baseUrl,
         headers: { 'Content-Type': 'application/json' },
-        withCredentials: true,
       })
     }
     this.accessToken = accessToken